+++ lib/utils/api_client.dart
@@ -1,12 +1,12 @@
 import 'dart:convert';
 
 import 'package:asgardeo_flutter_app/providers/user.dart';
-import 'package:asgardeo_flutter_app/utils/Auth.dart';
+import 'package:asgardeo_flutter_app/utils/auth_client.dart';
 import 'package:asgardeo_flutter_app/utils/util.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../configs/configs.dart';
-import '../configs/endPointUrls.dart';
+import '../configs/end_point_urls.dart';
 import '../constants.dart';
 import '../providers/page.dart';
 import '../providers/user_session.dart';
